+++ src/Engine/RenderTarget.cpp
@@ -1,7 +1,7 @@
 #include "RenderTarget.hpp"
 #include <Common/Log.hpp>
 
-#include "Shader/Shader.hpp"
+#include <Renderer/Shader/Shader.hpp>
 #include "Shader/ShaderProgram.hpp"
 
 #include "Geometry/Rectangle.hpp"
