+++ src/Engine/Shader/ShaderProgram.cpp
@@ -1,6 +1,6 @@
 #include "ShaderProgram.hpp"
 
-#include "Shader.hpp"
+#include <Renderer/Shader/Shader.hpp>
 
 ShaderProgram::ShaderProgram(std::initializer_list<const Shader*> shaders) {
     shaderProgram = glCreateProgram();
